requirements.txt
google-cloud-storage==3.1.0
Pillow==11.1.0
numpy==1.26.4
# torch==2.2.0
torch
torchvision==0.17.0
tqdm==4.67.1
matplotlib==3.10.1
pandas==2.2.3
